GALCORE: refcount interrupt handling to prevent shutdown race
This fixes the following scenario.  First thread comes in with
a command queue, it takes the powersave mutex and then enables
interrupts if they are disabled, submits the command queue and
releases the powersave mutex.  Then the program doing the drawing
exits and releases the driver, which takes the powersave mutex
and issues a powerdown which disables irqs.  This can cause
an interrupts pending softirq crash.  Therefore we refcount all
the calls to enable interrupts and don't disable them until we are
back down to 0.