brcmfmac: only report if_add errors if acting on them
We always get an error message on boot because we add the main
interface then enable events and get the event the interface
was added so we try to add it again.  Only report the error if
we are going to try and destroy the existing interface to try
and continue.

Also move the Event ignored error to be a debug EVENT message
as it is just a NOOP that is handled by the caller.

(cherry picked from commit 2071da50ea0bd966cf821ed00dd842dd55552b43)
Signed-off-by: Matus Kral <matuskral@me.com>