net: fec: probe phy and autonegotiate on fec probe.
The phy autonegotiation can take up to 4 seconds to happen when
connected to a gigabit switch.  Instead of waiting for userspace
to open the network device and then probing the phy we probe
the phy immediately after the netdev is registered and start
the autonegiotiation process.  This means that as soon as userspace
is ready we have link and can configure the network accordingly.

(cherry picked from commit d8d108b9cfe71354f606313d7fe5cecdfa100281)
Signed-off-by: Matus Kral <matuskral@me.com>

Conflicts:
	drivers/net/ethernet/freescale/fec_main.c