arm: configs: update cbi/hb configs to use patched galcore driver
To make it easier to use galcore upstream and also accept and rebase
on top of newer freescale kernels I have stopped patching the fsl
provided galcore driver and started working with a copy of it that
lives in the drivers/gpu directory.

By default build cbi/hb kernels with our version of the driver.