video: mxcfb: Pad memory allocations for odd resolutions.
This should be properly handled in userspace but I have seen a couple
of media applications that are not properly handling resolutions that
aren't aligned to 8 pixels like the gpu and vpu prefer.  This bug
presented itself in memory corruption when displaying output on a monitor
that has a 1600x900 resolution.  I will chase after userspace, but since
userspace tools that I can't change expect this alignment I will allocate
memory that is aligned to provide enough padding so they won't corrupt
system memory if used carelessly.