video: mxcfb: blank/unblank to save power
This patch does two things.

1)  It makes sure the display is unblanked when first connected.
This gives us a good user experience where if the framebuffer is
blanked and then a device is plugged in nothing turns on.

2)  After the hdmi device is registered we tell the IPU to blank
the display, then on subsequent connect/disconnect events we
unblank/blank the screens as needed.  This makes sure that the
ipu is always disabled if we aren't using it and if you don't have
a monitor attached saves about 80mA's of constant power usage.

Overall this provides a significnat power savings for headless
systems, and also doesn't block the bus-frequency driver from
kicking in if no display is attached.

(cherry picked from commit 056b42c28cd2c9123f5279f4f5a1d105f91f0853)
Signed-off-by: Matus Kral <matuskral@me.com>