gpu: galcore: Second try at cleaning up idle pm
The last commit could hit a sequence of events that led to the gpu
not powering off at idle as expected.  After inspecting a bit more
I realized we were doing a lot of uncessary work.  Rather than
get the pm mutex, query the hardware for idle, notify the os,
which then sets the pm_state, that gets the pm_mutex, or fails,
and then does all the other checks.  I move the idle query into
the pm_state setting and which means we only need to get the
mutex once, and if that fails we happily move along because
something else is changing the state so even if we are idle we
may not be soon.

On top of that I made the case where we request idle and then
fail to acquire the mutex an error.  This is and should be a noop.
Hey, Let's idle/suspend the gpu...oh we can't get the mutex, well
I guess something else is doing something so we can't idle/suspend,
lets bail out so we don't deadlock.

Don't send errors to the logs that aren't errors.

(cherry picked from commit a9fbd5d6a35699b4cc01b604559a9164b74eb053)