gpu: galcore: Relax locking on the command commit atom
This is supposed to be an atomic action of whether we have a
commit queued that should keep power management from stopping it
from getting to hardware. It makes sense to hold the PM mutex if
we are adding a commit, however if we are done with a commit and
decrementing then it is more efficient to change this value even
if the PM mutex is held so we may be able to catch the process
midway through.