gpu: galcore: Try harder to get PM mutex.
Upstream-commit: cd688d1c6c10f0723e31ab37ceb87f259386447c

I have cleaned up unnecessary pm changes so now we know we are
requesting a PM change only when we know we need it.  We can
then afford to try a bit harder here and waste some time because
it will only be a rare occasion we hit here, or the device should
be going into idle.