gpu: galcore: fix race in user signalling
Upstream-commit: 871aafd6aa6c61eb7147aa90a4580a878d2640c1

It was possible that betwen calling gckOS_MapSignal and
gckOS_Signal the signal could be destroyed by another thread.
In order to minimize the chance of this happening but keep
the current api mostly intact I have split out the core work
of gckOS_Signal, and gckOS_Destroy into Ex versions of the
functions that do the core work, while the previous existing
versions are the locking versions.  This allows me to call
the new functions while a lock is held in gckOS_UserSignal during
the entire Map, signal, Unmap flow.