video: mxc_edid: Adjust pixel timings
Our approach for fractional mode support requires the pixel timings
in the CEA mode table to describe the non-fractional modes. This
patch sets this for the modes that were defined otherwise.
Also it tweaks the rounding when calculating the real pixel clock
from the pixel time specified in the table.

Signed-off-by: Rudi <r.ihle@s-t.de>