phy: generate swphy registers on the fly
Upstream-commit: 703132a5ad553c7c70bef05126517ab4836334d3

Generate software phy registers as and when requested, rather than
duplicating the state in fixed_phy.  This allows us to eliminate
the duplicate storage of of the same data, which is only different
in format.

As fixed_phy_update_regs() no longer updates register state, rename
it to fixed_phy_update().

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>