Subject: [PATCH v2 01/24] mmc: core: shut up "voltage-ranges unspecified" pr_info()
Newsgroups: gmane.linux.kernel.mmc
Date: 2015-12-21 11:40:26 GMT (1 week, 6 days, 5 hours and 54 minutes ago)
Each time a driver such as sdhci-esdhc-imx is probed, we get a info
printk complaining that the DT voltage-ranges property has not been
specified.

However, the DT binding specifically says that the voltage-ranges
property is optional.  That means we should not be complaining that
DT hasn't specified this property: by indicating that it's optional,
it is valid not to have the property in DT.

Silence the warning if the property is missing.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>