Subject: [PATCH v2 17/24] mmc: sdhci: plug DMA mapping leak on error
Newsgroups: gmane.linux.kernel.mmc
Date: 2015-12-21 11:41:48 GMT (1 week, 6 days, 6 hours and 7 minutes ago)
If we terminate a command early, we fail to properly clean up the DMA
mappings for the data part of the request.  Move this clean up to the
tasklet, which is the common path for finishing a request so we always
clean up after ourselves.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>