Subject: [PATCH v2 21/24] mmc: sdhci: fix data timeout (part 2)
Newsgroups: gmane.linux.kernel.mmc
Date: 2015-12-21 11:42:09 GMT (1 week, 6 days, 6 hours and 8 minutes ago)
The calculation for the timeout based on the number of card clocks is
incorrect.  The calculation assumed:

    timeout in microseconds = clock cycles / clock in Hz

which is clearly a several orders of magnitude wrong.  Fix this by
multiplying the clock cycles by 1000000 prior to dividing by the Hz
based clock.  Also, as per part 1, ensure that the division rounds
up.

As this needs 64-bit math via do_div(), avoid it if the clock cycles
is zero.

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>